Fix padding-oracle-attack.md
The main problem is that the sentence 「如果 P 的最后一个字节是 i 的话，那么最后的 padding 就是 0x01」
is wrong, actually it should be 「如果 D_k(Y) 的最后一个字节是 i 的话，那么最后的 padding 就是 0x01」。

Instead of modifying it, I adjust the context to make it more accurate and straightforward. Also I
take the aforementioned CBC definition to better illustrate the process.

Moreover, there are several minor fixes are made:
- Fix notations to distinguish the crafted ones from the origin ones, such as P' and P, F' and F.
- The byte range of padding characters should be 1 to n instead of 0 to n.
- Wrap symbols with LaTeX '$' character to improve readability.